backend/app/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-cors
flask-limiter
psycopg2-binary
python-dotenv
requests
